+++ activityService/src/main/java/com/wefit/activityService/repository/ActivityRepository.java
@@ -1,9 +1,8 @@
-package com.wefit.activityService.repositories;
+package com.wefit.activityService.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
 import com.wefit.activityService.entities.Activity;
+import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ActivityRepository extends MongoRepository<Activity, String> {